backend/requirements.txt
Flask==1.1.2
uWSGI==2.0.19.1
mysql-connector==2.2.9
SQLAlchemy==1.3.17
Flask-Migrate==2.5.3
mysqlclient==2.0.1
Flask-Script==2.0.6
Flask-RESTful==0.3.8
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1



